pyproject.toml
[tool.poetry]
name = "hdb-chatbot"
version = "0.1.0"
description = "Chatbot for HDB information"
authors = ["Joanna Khek"]
readme = "README.md"
packages = [{include = "llm_model", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
pandas = ">=1.3.5"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
kaggle = "^1.5.16"
pyyaml = "^6.0.1"
numpy = "^1.21.4"
torch = {version = "^2.2.2+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.17.2+cu118", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.2.2+cu118", source = "pytorch-gpu-src"}
torchinfo = "^1.8.0"
scikit-learn = "^1.3.2"
torchmetrics = "^1.2.1"
tiktoken = "0.7"
pymupdf = "^1.24.3"
spacy = "^3.7.4"
sentence-transformers = "^2.7.0"
accelerate = "^0.30.1"
bitsandbytes = "^0.43.1"
packaging = "^24.0"
ninja = "^1.11.1.1"
transformers = "^4.40.2"
selenium = "^4.21.0"
streamlit = "^1.34.0"
beautifulsoup4 = "^4.12.3"
hydra-core = "^1.3.2"
langchain = "^0.2.5"
langchain-chroma = "^0.1.1"
openai = "^1.34.0"
langchain-openai = "^0.1.8"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.25.2"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"